requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
scipy
python-dotenv
requests
pytest